move decorator class to decorators
